test: Auctions
