+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -1287,85 +1287,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("HomeViewSectionSales", () => {
-    it("returns correct data", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-auctions") {
-              __typename
-              component {
-                title
-                behaviors {
-                  viewAll {
-                    buttonText
-                    href
-                    ownerType
-                  }
-                }
-              }
-
-              ... on HomeViewSectionSales {
-                salesConnection(first: 2) {
-                  edges {
-                    node {
-                      name
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const sales = [
-        {
-          name: "Auction 1",
-        },
-        {
-          name: "Auction 2",
-        },
-      ]
-
-      const context = {
-        salesLoader: jest.fn().mockResolvedValue(sales),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionSales",
-          "component": Object {
-            "behaviors": Object {
-              "viewAll": Object {
-                "buttonText": "Browse All Auctions",
-                "href": "/auctions",
-                "ownerType": "auctions",
-              },
-            },
-            "title": "Auctions",
-          },
-          "salesConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "name": "Auction 1",
-                },
-              },
-              Object {
-                "node": Object {
-                  "name": "Auction 2",
-                },
-              },
-            ],
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionTasks", () => {
     beforeAll(() => {
       mockIsFeatureFlagEnabled.mockImplementation((flag: string) => {
